native_src/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "transcription-suite-native"
version = "2.0.0"
description = "TranscriptionSuite native development environment"
readme = "../README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "homelab-00" }
]

# Same dependencies as server + desktop extras for native development
dependencies = [
    # Web / API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    
    # Core ML / Transcription
    "faster-whisper>=1.0.0",
    "ctranslate2>=4.0.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    
    # Audio Processing
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.0",
    "webrtcvad>=2.0.10",
    
    # Database
    "aiosqlite>=0.19.0",
    "sqlalchemy>=2.0.0",
    
    # Utilities
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "filelock>=3.13.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
# Speaker diarization
diarization = [
    "pyannote.audio>=3.1.0",
]

# Desktop/native development (system tray, audio recording)
desktop = [
    "PyQt6>=6.5.0",
    "pyaudio>=0.2.13",
    "pyperclip>=1.8.0",
]

# Full native development setup
full = [
    "pyannote.audio>=3.1.0",
    "PyQt6>=6.5.0",
    "pyaudio>=0.2.13",
    "pyperclip>=1.8.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["MAIN", "DIARIZATION"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]
